- Implemented Issue, Mint, Burn, with tests pased on testnet
