+++ README.md
@@ -306,12 +306,13 @@ is_valid = wallet.verify_signature(msg, signature)
 from binancechain.enums import Ordertype, Side, Timeinforce, Votes
 
 #if client is passed in , testnet arg will be ignored
+transaction = Transaction(wallet=wallet, client=client)
 
-transaction = Transaction(wallet=wallet, client=None,testnet=False)
+transfer = await transaction.transfer(
+    to_address=wallet_two.get_address(), symbol="BNB", amount=0.1
+)
 
-transfer = await transaction.transfer(to_address, symbol, amount)
-
-broadcast_info = await transaction.create_new_order(
+new_order_txid = await transaction.create_new_order(
     symbol="binance_pair",
     side=Side.BUY,
     ordertype=Ordertype.LIMIT,
@@ -320,15 +321,19 @@ broadcast_info = await transaction.create_new_order(
     timeInForce=Timeinforce.GTE,
 )
 
-broadcast_info = await transaction.cancel_order(symbol="pair", refid)
+cancel_order_txid = await transaction.cancel_order(symbol="pair", refid="")
 
-broadcast_info = await transaction.freeze_token(symbol="token", amount)
+freeze_token_txid = await transaction.freeze_token(symbol="token", amount=1)
 
-broadcast_info = await transaction.unfreeze_token(symbol="token", amount)
+unfreeze_token_txid = await transaction.unfreeze_token(symbol="token", amount=1)
 
-broadcast_info = await transaction.vote(
-    proposal_id="", option=VOTES.Yes
-)
+vote_txid = await transaction.vote(proposal_id, option=Votes.YES)  # only validator can vote
+
+issue_token_txid = await transaction.issue_token(symbol, name, supply, mintable)
+
+mint_token_txid = await transaction.mint_token(symbol, amount)
+
+burn_token_txid = await transaction.burn_token(symbol, amount)
 ```
 ### Create Transaction Message. This message can be signed and broadcast somewhere else
 
@@ -341,10 +346,8 @@ limit_buy_transaction = await Transaction.new_order_transaction(
       price=1,
       quantity=1,
       timeInForce=Timeinforce.GTE,
-      account_number= None,
-      sequence= None,
-      client= None,
-      testnet = False, # will be ignored if client is passed in
+      testnet=True,
+      client=None,
   )
 
 limit_sell_transaction = await Transaction.new_order_transaction(
@@ -355,40 +358,36 @@ limit_sell_transaction = await Transaction.new_order_transaction(
     price=1,
     quantity=1,
     timeInForce=Timeinforce.GTE,
-    account_number= None,
-    sequence= None,
-    client= None,
-    testnet = False, # will be ignored if client is passed in
+    testnet=True,
+    client=None,
 )
 
 cancel_order_transaction = await Transaction.cancel_order(
-    address="owner_address",
-    symbol="pair",
-    refid="",
-    account_number= None,
-    sequence= None,
-    client= None,
-    testnet = False, # will be ignored if client is passed in
+    address="owner_address", symbol="pair", refid="", testnet=True, client=None
 )
 
 freeze_token_transaction = await Transaction.freeze_token(
-    address="ownder_address",
-    symbol="token",
-    amount=1,
-    account_number= None,
-    sequence= None,
-    client= None,
-    testnet = False, # will be ignored if client is passed in
+    address="ownder_address", symbol="token", amount=1, testnet=True, client=None
 )
 
 unfreeze_token_tranasaction = await Transaction.unfreeze_token_transaction(
-    address="ownder_address",
-    symbol="token",
-    amount=1,
-    account_number= None,
-    sequence= None,
-    client= None,
-    testnet = False, # will be ignored if client is passed in
+    address="ownder_address", symbol="token", amount=1, testnet=True, client=None
+)
+
+vote_transaction = await Transaction.vote_transaction(
+    voter, proposal_id, option=Votes.YES, client=None, testnet=True
+)
+
+issue_token_transaction = await Transaction.issue_token_transaction(
+    owner, name, symbol, sypply, mintable, client=None, testnet=True
+)
+
+mint_token_transaction = await Transaction.mint_token_transaction(
+    owner, symbol, amount, client=None, testnet=True
+)
+
+burn_token_transaction = Transaction.burn_token_transaction(
+    owner, symbol, amount, client=None, testnet=True
 )
 
 """ Get transaction message to sign"""
