+++ examples/transaction_examples.py
@@ -13,7 +13,7 @@ async def transaction_example():
         to_address=wallet_two.get_address(), symbol="BNB", amount=0.1
     )
 
-    broadcast_info = await transaction.create_new_order(
+    new_order_txid = await transaction.create_new_order(
         symbol="binance_pair",
         side=Side.BUY,
         ordertype=Ordertype.LIMIT,
@@ -22,16 +22,21 @@ async def transaction_example():
         timeInForce=Timeinforce.GTE,
     )
 
-    broadcast_info = await transaction.cancel_order(symbol="pair", refid="")
+    cancel_order_txid = await transaction.cancel_order(symbol="pair", refid="")
 
-    broadcast_info = await transaction.freeze_token(symbol="token", amount=1)
+    freeze_token_txid = await transaction.freeze_token(symbol="token", amount=1)
 
-    broadcast_info = await transaction.unfreeze_token(symbol="token", amount=1)
+    unfreeze_token_txid = await transaction.unfreeze_token(symbol="token", amount=1)
 
-    broadcast_info = await transaction.vote(
+    vote_txid = await transaction.vote(
         proposal_id="", option=Votes.YES
     )  # only validator can vote
 
+    issue_token_txid = await transaction.issue_token(symbol, name, supply, mintable)
+
+    mint_token_txid = await transaction.mint_token(symbol, amount)
+
+    burn_token_txid = await transaction.burn_token(symbol, amount)
     """
     Create Unsigned Transaction, return transaction with message to sign and broadcast somewhere else
     """
@@ -47,6 +52,7 @@ async def transaction_example():
         quantity=1,
         timeInForce=Timeinforce.GTE,
         testnet=True,
+        client=None,
     )
 
     limit_sell_transaction = await Transaction.new_order_transaction(
@@ -58,21 +64,38 @@ async def transaction_example():
         quantity=1,
         timeInForce=Timeinforce.GTE,
         testnet=True,
+        client=None,
     )
 
     cancel_order_transaction = await Transaction.cancel_order(
-        address="owner_address", symbol="pair", refid="", testnet=True
+        address="owner_address", symbol="pair", refid="", testnet=True, client=None
     )
 
     freeze_token_transaction = await Transaction.freeze_token(
-        address="ownder_address", symbol="token", amount=1, testnet=True
+        address="ownder_address", symbol="token", amount=1, testnet=True, client=None
     )
 
     unfreeze_token_tranasaction = await Transaction.unfreeze_token_transaction(
-        address="ownder_address", symbol="token", amount=1, testnet=True
+        address="ownder_address", symbol="token", amount=1, testnet=True, client=None
+    )
+
+    vote_transaction = await Transaction.vote_transaction(
+        voter, proposal_id, option=Votes.YES, client=None, testnet=True
+    )
+
+    issue_token_transaction = await Transaction.issue_token_transaction(
+        owner, name, symbol, sypply, mintable, client=None, testnet=True
+    )
+
+    mint_token_transaction = await Transaction.mint_token_transaction(
+        owner, symbol, amount, client=None, testnet=True
+    )
+
+    burn_token_transaction = Transaction.burn_token_transaction(
+        owner, symbol, amount, client=None, testnet=True
     )
 
     """
-    Get information from transaction object
+    Get Sign Message to sign with wallet
     """
-    sign_message_bytes_format = Limit_Buy_Transaction.get_sign_message()
+    sign_message_bytes_format = limit_buy_transaction.get_sign_message()
